created new singleWhiteboard reducer
