+++ client/components/Attendees.js
@@ -4,15 +4,10 @@ import { withRouter } from 'react-router'
 
 export class Attendees extends Component {
   render() {
-    //getting whiteboard id
-    const whiteboardId = this.props.match.params.id
-    //finding the whiteboard in array of whiteboard returned from state
-    const foundWhiteboard = this.props.whiteboard.find((whiteboard) => {
-      return whiteboardId == whiteboard.id
-    })
-    //getting users from that found whiteboard
+
+    const foundWhiteboard = this.props.whiteboard
+
     const { users } = foundWhiteboard
-    console.log('FOUND---', foundWhiteboard)
     return (
       <div>
         <p>Host: {foundWhiteboard.host}</p>
@@ -26,7 +21,7 @@ export class Attendees extends Component {
 
 const mapState = state => {
   return {
-    whiteboard: state.whiteboard,
+    whiteboard: state.singleWhiteboard,
     user: state.user
   }
 }
